Adds handling of @tricorder annotations
This adds basic handling of @tricorder annotations. This allows you
to put @tricorder coversMethodAcceptsStringValues to denote that
a test exists that verifies that the method only accepts string
values, or @tricorder coversMethodReturnsVoidValues to denote that
a test exists that verifies that the method doesn't return anything.

This gives a user the ability to run Tricorder in a CI environment
and get suggestions on what to test next, and then mark it as
"complete" so as not to see it again on the next run.

It uses a simple preg_match, but requires that the annotation match
@tricorder coversMethodReturns(.*?)Values and
@tricorder coversMethodAccepts(.*?)Values.